djangobackend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
django==3.1.3
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
httpie==2.3.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itypes==1.2.0
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
openapi-codec==1.3.2
pygments==2.7.2; python_version >= '3.5'
pyjwt==1.7.1
pysocks==1.7.1
pytz==2020.4
pyyaml==5.3.1
requests-toolbelt==0.9.1
requests[socks]==2.25.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
simplejson==3.17.2; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
uritemplate==3.0.1
urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
